+++ lib/Data/Time/Format/Parse.hs
@@ -7,9 +7,6 @@ module Data.Time.Format.Parse
     , parseTimeOrError
     , readSTime
     , readPTime
-    , parseTime
-    , readTime
-    , readsTime
     , ParseTime()
     -- * Locale
     , module Data.Time.Format.Locale
@@ -127,42 +124,6 @@ readPOnlyTime ::
     -> String -- ^ Format string
     -> ReadP t
 readPOnlyTime = readPOnlyTime' Proxy
-
-{-# DEPRECATED
-parseTime "use \"parseTimeM True\" instead"
- #-}
-
-parseTime ::
-       ParseTime t
-    => TimeLocale -- ^ Time locale.
-    -> String -- ^ Format string.
-    -> String -- ^ Input string.
-    -> Maybe t -- ^ The time value, or 'Nothing' if the input could
-                        -- not be parsed using the given format.
-parseTime = parseTimeM True
-
-{-# DEPRECATED
-readTime "use \"parseTimeOrError True\" instead"
- #-}
-
-readTime ::
-       ParseTime t
-    => TimeLocale -- ^ Time locale.
-    -> String -- ^ Format string.
-    -> String -- ^ Input string.
-    -> t -- ^ The time value.
-readTime = parseTimeOrError True
-
-{-# DEPRECATED
-readsTime "use \"readSTime True\" instead"
- #-}
-
-readsTime ::
-       ParseTime t
-    => TimeLocale -- ^ Time locale.
-    -> String -- ^ Format string
-    -> ReadS t
-readsTime = readSTime True
 
 -- * Read instances for time package types
 instance Read Day where